Updated CreateAccountPanel
Doctors.csv uploads to the server.
All the .jar files needs to be added to your build path for the imports
to work.
Since we have a private repo - we should be good concerning secret keys,
etc.
There are still some issues - I just need to upload doctors.csv first
before checking if it exists, and appending - that way it doesn't upload
a copy that's not up-to-date, if that makes sense.